Add VCWithOverlappingElementsSpec file.
